Enhance task configuration avoidance tests
Better test names.

Allow 'clean' task configuration as Gradle core force configures it in a lot of cases.

Rework how we detect other tasks configuration when scheduling the plugin under test task by comparing the set of scheduled tasks with the set of configured tasks. Also allow test implementations to provide an extra set of configured tasks for progressive adoption of the test.